==> Notebooks/4_compare_peaks_lists.ipynb <==
import pandas as pd
import os

wd_dir = '/beegfs/scratch/ric.broccoli/kubacki.michal/SRF_CUTandTAG'
os.chdir(wd_dir)

def find_intersection(file1_path, file2_path, gene_list=None):
    """
    Find intersection between two CSV files based on gene, Endogenous_Promoter, 
    and Exogenous_Promoter columns.
    """
    # Read the CSV files
    df1 = pd.read_csv(file1_path)
    df2 = pd.read_csv(file2_path)
    
    if gene_list is not None:
        # Filter df1 to only include genes from gene_list
        df1 = df1[df1['gene'].isin(gene_list)]
    
    # Check for duplicates in the matching columns
    columns_to_compare = ['gene', 'Endogenous_Promoter', 'Exogenous_Promoter']
    
    # Print duplicate information
    print("\nChecking for duplicates in file 1:")
    duplicates1 = df1[df1.duplicated(subset=columns_to_compare, keep=False)]
    print(f"Number of duplicate rows in file 1: {len(duplicates1)}")
    if len(duplicates1) > 0:
        print("\nExample duplicates from file 1:")
        print(duplicates1.head())
    
    print("\nChecking for duplicates in file 2:")
    duplicates2 = df2[df2.duplicated(subset=columns_to_compare, keep=False)]
    print(f"Number of duplicate rows in file 2: {len(duplicates2)}")
    if len(duplicates2) > 0:
        print("\nExample duplicates from file 2:")
        print(duplicates2.head())
    
    # Ensure boolean values are consistent
    for df in [df1, df2]:
        df['Endogenous_Promoter'] = df['Endogenous_Promoter'].astype(bool)
        df['Exogenous_Promoter'] = df['Exogenous_Promoter'].astype(bool)
    
    # Add this before the merge
    df1 = df1.drop_duplicates(subset=columns_to_compare)
    df2 = df2.drop_duplicates(subset=columns_to_compare)
    
    # Merge the dataframes on the specified columns
    merged_df = pd.merge(df1, 
                        df2,
                        on=columns_to_compare,
                        suffixes=('_file1', '_file2'))
    
    # Create a summary of the results
    total_rows_file1 = len(df1)
    total_rows_file2 = len(df2)
    matching_rows = len(merged_df)
    
    print(f"\nTotal rows in file 1: {total_rows_file1}")
    print(f"Total rows in file 2: {total_rows_file2}")
    print(f"Number of matching rows: {matching_rows}")
    
    return merged_df

## Neuron disregulated genes

# Load gene lists from CSV files
upregulated = pd.read_csv('custom_pipeline/DATA/neuron_upregulated_genes.csv')['gene'].tolist()
downregulated = pd.read_csv('custom_pipeline/DATA/neuron_downregulated_genes.csv')['gene'].tolist() 

disregulated = upregulated + downregulated

file1_path = 'custom_pipeline/results/Neuron_peak_analysis_annotated_clean.csv'
file2_path = 'custom_pipeline/DATA/allgenes_NEU_total.csv'

# %%script false --no-raise-error
# # Find intersection
# result = find_intersection(file1_path, file2_path)

# # Display the first few rows of the intersection
# result.head()

# # save the results
# # path_to_save = 'custom_pipeline/results'
# # result.to_csv(f'{path_to_save}/matches_Neuron_file2.csv', index=False)

# Find intersection
result = find_intersection(file1_path, file2_path, disregulated)

result.head()

result_up = find_intersection(file1_path, file2_path, upregulated)

result_down = find_intersection(file1_path, file2_path, downregulated)

## NSC disregulated genes

# Load gene lists from CSV files
upregulated = pd.read_csv('custom_pipeline/DATA/nsc_upregulated_genes.csv')['gene'].tolist()
downregulated = pd.read_csv('custom_pipeline/DATA/nsc_downregulated_genes.csv')['gene'].tolist() 

disregulated = upregulated + downregulated

file1_path = 'custom_pipeline/results/NSC_peak_analysis_annotated_clean.csv'
file2_path = 'custom_pipeline/DATA/allgenes_NSC_total.csv'

# %%script false --no-raise-error
# # Find intersection
# result = find_intersection(file1_path, file2_path)

# # Display the first few rows of the intersection
# result.head()

# # save the results
# # path_to_save = 'custom_pipeline/results'
# # result.to_csv(f'{path_to_save}/matches_NSC_file2.csv', index=False)

# Find intersection
result = find_intersection(file1_path, file2_path, disregulated)

result.head()

result_up = find_intersection(file1_path, file2_path, upregulated)

result_down = find_intersection(file1_path, file2_path, downregulated)

   location              Endogenous_Promoter Exogenous_Promoter   chrom  coords               start      end         gene     baseMean      log2FoldChange   Direction
0  chr1:55405521-55406190        True                True         chr1   55405521-55406190    55405521   55406190    Plcl1    361.918949    -0.726832        DOWN
1  chr1:62706498-62707027        True                True         chr1   62706498-62707027    62706498   62707027    Nrp2     1460.062193   -0.541600        DOWN
2  chr1:89070006-89070435        True                True         chr1   89070006-89070435    89070006   89070435    Sh3bp4   630.231250    -0.553894        DOWN
3  chr1:91413167-91413489        True                True         chr1   91413167-91413489    91413167   91413489    Hes6     429.655435    0.631325         UP
4  chr1:106581988-106582206      True                True         chr1   106581988-106582206  106581988  106582206   Bcl2     456.720686    -0.819810        DOWN
